+++ lizard_progress/urls.py
@@ -16,15 +16,11 @@ from lizard_map.views import AppView
 from lizard_progress.forms import CalculateForm
 from lizard_progress.forms import ContractorChoiceForm
 from lizard_progress.forms import ContractorForm
-from lizard_progress.forms import ExistingUserForm
 from lizard_progress.forms import MeasurementTypeForm
-from lizard_progress.forms import NewUserForm
 from lizard_progress.forms import ProjectChoiceForm
 from lizard_progress.forms import ProjectForm
 from lizard_progress.forms import ShapefileForm
-from lizard_progress.forms import existing_user_condition
 from lizard_progress.forms import needs_shapefile_condition
-from lizard_progress.forms import new_user_condition
 
 from lizard_progress.views import ActivitiesWizard
 from lizard_progress.views import ComparisonPopupView
@@ -73,10 +69,7 @@ urlpatterns = patterns(
         name='lizard_progress_newproject'),
     url('^admin/contractors/new/$',
         login_required(ContractorWizard.as_view(
-            [ContractorForm, ExistingUserForm, NewUserForm],
-        condition_dict={
-                '1': existing_user_condition,
-                '2': new_user_condition})),
+            [ContractorForm])),
         name='lizard_progress_newcontractor'),
     url('^admin/activities/new/$',
         login_required(ActivitiesWizard.as_view(